requirements.txt
uniswap-python
web3
python-dotenv